refactor: move 'parseRepositoryErrors' method to 'mongoose' service
