+++ src/services/mongoose/accounts.ts
@@ -9,7 +9,7 @@ import {
 } from "@domain/errors"
 
 import { User } from "@services/mongoose/schema"
-import { parseRepositoryError } from "@services/shared/repository"
+import { parseRepositoryError } from "@services/mongoose/utils"
 
 import { toObjectId, fromObjectId } from "./utils"
