+++ src/services/mongoose/utils.ts
@@ -1,3 +1,4 @@
+import { CantConnectToMongoDbError, UnknownRepositoryError } from "@domain/errors"
 import { Types } from "mongoose"
 
 export const toObjectId = <T extends string>(id: T): Types.ObjectId => {
@@ -6,4 +7,14 @@ export const toObjectId = <T extends string>(id: T): Types.ObjectId => {
 
 export const fromObjectId = <T extends string>(id: Types.ObjectId | string): T => {
   return String(id) as T
+}
+
+export const parseRepositoryError = (err: Error) => {
+  switch (true) {
+    case err.message.includes("MongoNetworkTimeoutError: connection timed out"):
+    case err.message.includes("MongooseServerSelectionError: connection timed out"):
+      return new CantConnectToMongoDbError()
+    default:
+      return new UnknownRepositoryError(err.message)
+  }
 }
