+++ src/services/mongoose/wallet-on-chain-addresses.ts
@@ -1,7 +1,7 @@
 import { CouldNotFindError, PersistError, RepositoryError } from "@domain/errors"
 import { baseLogger } from "@services/logger"
 import { Wallet } from "@services/mongoose/schema"
-import { parseRepositoryError } from "@services/shared/repository"
+import { parseRepositoryError } from "@services/mongoose/utils"
 
 export const WalletOnChainAddressesRepository = (): IWalletOnChainAddressesRepository => {
   const persistNew = async ({
