+++ src/services/payment-flow/index.ts
@@ -8,7 +8,7 @@ import {
 import { InvalidLightningPaymentFlowStateError, PaymentFlow } from "@domain/payments"
 import { paymentAmountFromNumber, WalletCurrency } from "@domain/shared"
 import { safeBigInt } from "@domain/shared/safe"
-import { parseRepositoryError } from "@services/shared/repository"
+import { parseRepositoryError } from "@services/mongoose/utils"
 import { elapsedSinceTimestamp } from "@utils"
 
 import { PaymentFlowState } from "./schema"
